+++ src/main/java/com/demonwav/statcraft/commands/sc/SCMined.java
@@ -44,7 +44,7 @@ public class SCMined extends SCTemplate implements CustomResponse {
     }
 
     @Override
-    public String serverStatListResponse(int num, List<String> args, Connection connection) {
+    public String serverStatListResponse(long num, List<String> args, Connection connection) {
         return null;
     }
